Install C objects with a user-supplied prefix.
All this work came about because I was getting a weird issue with
cbindgen when using mwalib as a crate dependency. But, this is something
I've wanted to investigate for a while.

The old compilation behaviour was to put mwalib.h in ./include/, which
is against cargo's wishes. On top of this, if a user was building from
source, they would have to take the header from this custom spot and the
libraries from ./target/release. This commit uses cargo-make to put the
C libraries and the header in a user-specified prefix. This allows
cbindgen to be an optional dependency, which is great for Rust users.

cbindgen was in dev-dependencies previously (probably an oversight).

I also tweaked the README's installation instructions.

This commit also introduces .cargo/config.toml, which instructs cargo to
build against the local CPU for maximum performance.